Fixed videoP2P hang issue
BZ: 168381

When network is disconnected, videoP2P logged out, which destroyed
GCallClient where it would destroy the mutex.  When videoP2P logged
in again, it did not re-initialize the mutex, so any lock to the mutex
will be blocked, which caused the deadlock.  This patch will make sure
the mutex will initialize after it was being destroyed.

Change-Id: Ia0edb649dbbc04f06fda8ebcc3e221c32544a67b
Signed-off-by: Khanh V. Nguyen <khanh.v.nguyen@intel.com>