Make orientation change smooth, when device is rotated
BZ: 166717

When the phone was rotated in VideoP2P app, the image would flip
90 degrees before continuing to rotate. Fixed the threshold of
rotation to a larger degree, so that rotation looks smoother.

Another issue noticed was that If the call is started in landscape
mode, the image orientation is still in portrait mode because of
improper init values. Fixed the init values to resolve this problem.

Change-Id: I2364ac2b4705cb3ac751733f6747f25533f35c87
Signed-off-by: Sushant Painaik <spainaix.painaik@intel.com>